+++ examples/datasets/sample_car_data.csv
@@ -1,0 +1,12 @@
+time,position,velocity
+0,0,0
+1,2,2
+2,8,6
+3,18,10
+4,32,14
+5,50,18
+6,72,22
+7,98,26
+8,128,30
+9,162,34
+10,200,38
