+++ examples/datasets/sample_physics_data.csv
@@ -1,0 +1,12 @@
+time,position,mass,temperature
+0,0,1.5,20
+1,2,1.5,22
+2,8,1.5,25
+3,18,1.5,28
+4,32,1.5,30
+5,50,1.5,32
+6,72,1.5,35
+7,98,1.5,37
+8,128,1.5,40
+9,162,1.5,42
+10,200,1.5,45
